move option_as_ref_deref to its own module
